fix(hub): settings freshness hydrates from custody, not just create
The /settings + tray freshness check (providerBaseFreshness) read only local
prepared-state, which is empty on a control box — bakes live in custody and were
hydrated only at create time. So every provider showed 'needs baking' even when
custody held a matching record. Hydrate from custody in the freshness path too
(same fingerprint-match-wins policy, reused via the extracted
./prepared-hydrate helper), so /settings reflects shared bakes. No-op on a local
hub or when custody has no matching record.
